+++ sampledb/frontend/objects_forms.py
@@ -103,3 +103,7 @@ class ObjectPublicationForm(FlaskForm):
             field.data = simplify_doi(field.data)
         except InvalidDOIError:
             raise ValidationError('Please enter a valid DOI')
+
+
+class CopyPermissionsForm(FlaskForm):
+    object_id = SelectField(validators=[InputRequired()])